Fluff C3 seam 2: extract inline tracked changes to trackedChanges.ts
Verbatim move of the proposed-edit types, the strict/loose anchor
resolver (locateProposedEditInDoc), the strict apply path, the
accept/reject widget, and the decorations extension. DocumentRichEditor
re-exports locateProposedEditInDoc, DocumentProposedEdit and
ProposedEditLocation so DocumentDetail's import path is unchanged.
